requirements.txt
ipython
numpy
pyflakes
pysoundfile
https://github.com/20tab/mpg123-python/archive/master.zip
https://github.com/larsimmisch/pyalsaaudio/archive/master.zip
